Make the main controller importable
Merge `core/core.js` in `core/core.controller.js`, split defaults options next to their associated code and deprecate `Chart.type` (not sure what it was for).
