+++ src/core/core.defaults.js
@@ -2,7 +2,7 @@
 
 var helpers = require('../helpers/helpers.core');
 
-module.exports = {
+var defaults = {
 	/**
 	 * @private
 	 */
@@ -10,3 +10,15 @@ module.exports = {
 		return helpers.merge(this[scope] || (this[scope] = {}), values);
 	}
 };
+
+defaults._set('global', {
+	defaultColor: 'rgba(0,0,0,0.1)',
+	defaultFontColor: '#666',
+	defaultFontFamily: "'Helvetica Neue', 'Helvetica', 'Arial', sans-serif",
+	defaultFontSize: 12,
+	defaultFontStyle: 'normal',
+	defaultLineHeight: 1.2,
+	showLines: true
+});
+
+module.exports = defaults;
